Rename 'step73' to coffeehouse.
