+++ server/pom.xml
@@ -3,7 +3,7 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>com.google.step73</groupId>
+  <groupId>com.google.coffeehouse</groupId>
   <artifactId>coffeehouse</artifactId>
   <version>1</version>
   <packaging>war</packaging>
